Create When Can I Play, Dad?
I ran into an issue in my household where my young children would often disagree about whose turn it was to play on the video game console.  In order to alleviate the situation, I created a toy program that would randomize their names, and randomize play times for them, so it would remove me from the situation and I wouldn't have to play referee anymore!  I had a great time with this one. This was the first time I messed with the colors of the text in the output and I think it came out great.  I also included an easter egg in the code for my kids.  Can you find it?  :)